+++ src/main/java/com/phantomwing/rusticdelight/datagen/DataGenerators.java
@@ -19,7 +19,7 @@ import net.neoforged.neoforge.registries.NeoForgeRegistries;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 
-@EventBusSubscriber(bus = EventBusSubscriber.Bus.MOD, modid = RusticDelight.MOD_ID)
+@EventBusSubscriber(modid = RusticDelight.MOD_ID)
 public class DataGenerators {
     @SubscribeEvent
     public static void gatherData(GatherDataEvent event) {