+++ src/main/java/com/phantomwing/rusticdelight/food/FoodValues.java
@@ -3,7 +3,6 @@ package com.phantomwing.rusticdelight.food;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.food.FoodProperties;
-import vectorwing.farmersdelight.common.registry.ModEffects;
 
 import static vectorwing.farmersdelight.common.FoodValues.*;
 
@@ -135,7 +134,7 @@ public class FoodValues {
             .effect(() -> new MobEffectInstance(MobEffects.REGENERATION, 100, 0), 1.0F).build();
     public static final FoodProperties BELL_PEPPER_SOUP = (new FoodProperties.Builder())
             .nutrition(6).saturationModifier(0.6f)
-            .effect(() -> new MobEffectInstance(ModEffects.COMFORT, MEDIUM_DURATION, 0), 1.0F).build();
+            .effect(() -> nourishment(MEDIUM_DURATION), 1.0f).build();
 
     // Plated foods
     public static final FoodProperties BELL_PEPPER_PASTA = (new FoodProperties.Builder())