+++ src/main/java/com/phantomwing/rusticdelight/item/ModItems.java
@@ -18,6 +18,7 @@ import net.neoforged.neoforge.registries.DeferredItem;
 import net.neoforged.neoforge.registries.DeferredRegister;
 import vectorwing.farmersdelight.common.item.ConsumableItem;
 import vectorwing.farmersdelight.common.item.DrinkableItem;
+import vectorwing.farmersdelight.common.item.PlaceableItem;
 
 import java.util.LinkedHashSet;
 import java.util.function.Supplier;
@@ -133,12 +134,12 @@ public class ModItems {
             baseItem().food(FoodValues.COOKED_CALAMARI_SLICE)));
 
     // Pies & cakes
-    public static final DeferredItem<Item> SYRUP_CHEESECAKE = registerBlockWithTab(ModBlocks.SYRUP_CHEESECAKE, baseItem());
-    public static final DeferredItem<Item> SYRUP_CHEESECAKE_SLICE = registerWithTab("syrup_cheesecake_slice", () -> new Item(
+    public static final DeferredItem<Item> SYRUP_CHEESECAKE = registerPlaceableBlockWithTab(ModBlocks.SYRUP_CHEESECAKE, baseItem());
+    public static final DeferredItem<Item> SYRUP_CHEESECAKE_SLICE = registerWithTab("syrup_cheesecake_slice", () -> new ConsumableItem(
             baseItem().food(vectorwing.farmersdelight.common.FoodValues.PIE_SLICE)));
 
-    public static final DeferredItem<Item> CHERRY_BLOSSOM_CHEESECAKE = registerBlockWithTab(ModBlocks.CHERRY_BLOSSOM_CHEESECAKE, baseItem());
-    public static final DeferredItem<Item> CHERRY_BLOSSOM_CHEESECAKE_SLICE = registerWithTab("cherry_blossom_cheesecake_slice", () -> new Item(
+    public static final DeferredItem<Item> CHERRY_BLOSSOM_CHEESECAKE = registerPlaceableBlockWithTab(ModBlocks.CHERRY_BLOSSOM_CHEESECAKE, baseItem());
+    public static final DeferredItem<Item> CHERRY_BLOSSOM_CHEESECAKE_SLICE = registerWithTab("cherry_blossom_cheesecake_slice", () -> new ConsumableItem(
             baseItem().food(vectorwing.farmersdelight.common.FoodValues.PIE_SLICE)));
 
     // Cookies
@@ -156,22 +157,22 @@ public class ModItems {
             baseItem().food(FoodValues.FRUIT_BEIGNET), true));
 
     // Pancakes
-    public static final DeferredItem<Item> PANCAKES = registerBlockWithTab(ModBlocks.PANCAKES, bowlItem());
+    public static final DeferredItem<Item> PANCAKES = registerPlaceableBlockWithTab(ModBlocks.PANCAKES, bowlItem());
     public static final DeferredItem<Item> PANCAKE = registerWithTab("pancake", () -> new Item(
             baseItem().food(FoodValues.PANCAKE)));
-    public static final DeferredItem<Item> HONEY_PANCAKES = registerBlockWithTab(ModBlocks.HONEY_PANCAKES, bowlItem());
+    public static final DeferredItem<Item> HONEY_PANCAKES = registerPlaceableBlockWithTab(ModBlocks.HONEY_PANCAKES, bowlItem());
     public static final DeferredItem<Item> HONEY_PANCAKE = registerWithTab("honey_pancake", () -> new ConsumableItem(
             baseItem().food(FoodValues.HONEY_PANCAKE), true));
-    public static final DeferredItem<Item> CHOCOLATE_PANCAKES = registerBlockWithTab(ModBlocks.CHOCOLATE_PANCAKES, bowlItem());
+    public static final DeferredItem<Item> CHOCOLATE_PANCAKES = registerPlaceableBlockWithTab(ModBlocks.CHOCOLATE_PANCAKES, bowlItem());
     public static final DeferredItem<Item> CHOCOLATE_PANCAKE = registerWithTab("chocolate_pancake", () -> new ConsumableItem(
             baseItem().food(FoodValues.CHOCOLATE_PANCAKE), true));
-    public static final DeferredItem<Item> CHERRY_BLOSSOM_PANCAKES = registerBlockWithTab(ModBlocks.CHERRY_BLOSSOM_PANCAKES, bowlItem());
+    public static final DeferredItem<Item> CHERRY_BLOSSOM_PANCAKES = registerPlaceableBlockWithTab(ModBlocks.CHERRY_BLOSSOM_PANCAKES, bowlItem());
     public static final DeferredItem<Item> CHERRY_BLOSSOM_PANCAKE = registerWithTab("cherry_blossom_pancake", () -> new ConsumableItem(
             baseItem().food(FoodValues.CHERRY_BLOSSOM_PANCAKE), true));
-    public static final DeferredItem<Item> VEGETABLE_PANCAKES = registerBlockWithTab(ModBlocks.VEGETABLE_PANCAKES, bowlItem());
+    public static final DeferredItem<Item> VEGETABLE_PANCAKES = registerPlaceableBlockWithTab(ModBlocks.VEGETABLE_PANCAKES, bowlItem());
     public static final DeferredItem<Item> VEGETABLE_PANCAKE = registerWithTab("vegetable_pancake", () -> new ConsumableItem(
             baseItem().food(FoodValues.VEGETABLE_PANCAKE), true));
-    public static final DeferredItem<Item> PUMPKIN_PANCAKES = registerBlockWithTab(ModBlocks.PUMPKIN_PANCAKES, bowlItem());
+    public static final DeferredItem<Item> PUMPKIN_PANCAKES = registerPlaceableBlockWithTab(ModBlocks.PUMPKIN_PANCAKES, bowlItem());
     public static final DeferredItem<Item> PUMPKIN_PANCAKE = registerWithTab("pumpkin_pancake", () -> new ConsumableItem(
             baseItem().food(FoodValues.PUMPKIN_PANCAKE), true));
 
@@ -226,7 +227,7 @@ public class ModItems {
             bowlItem().food(FoodValues.COFFEE_BRAISED_BEEF), true));
 
     // Feasts
-    public static final DeferredItem<Item> RICE_ROLL_ROYALE = registerBlockWithTab(ModBlocks.RICE_ROLL_ROYALE, feastItem());
+    public static final DeferredItem<Item> RICE_ROLL_ROYALE = registerPlaceableBlockWithTab(ModBlocks.RICE_ROLL_ROYALE, feastItem());
 
     // Helper functions
     public static Item.Properties baseItem() {
@@ -258,6 +259,10 @@ public class ModItems {
 
     public static DeferredItem<Item> registerBlockWithTab(DeferredBlock<Block> block, Item.Properties properties) {
         return registerWithTab(block.getRegisteredName().replaceFirst(RusticDelight.MOD_ID + ":", ""), () -> new BlockItem(block.get(), properties));
+    }
+
+    public static DeferredItem<Item> registerPlaceableBlockWithTab(DeferredBlock<Block> block, Item.Properties properties) {
+        return registerWithTab(block.getRegisteredName().replaceFirst(RusticDelight.MOD_ID + ":", ""), () -> new PlaceableItem(block.get(), properties));
     }
 
     public static void register(IEventBus eventBus) {