+++ src/main/java/com/phantomwing/rusticdelight/tags/CommonTags.java
@@ -23,7 +23,7 @@ public class CommonTags {
     public static final TagKey<Item> FOODS_LEAFY_GREEN = commonItemTag("foods/leafy_green");
     public static final TagKey<Item> FOODS_DOUGH = commonItemTag("foods/dough");
     public static final TagKey<Item> FOODS_PASTA = commonItemTag("foods/pasta");
-    public static final TagKey<Item> FOODS_MILK = commonItemTag("foods/milk");
+    public static final TagKey<Item> DRINKS_MILK = commonItemTag("drinks/milk");
 
     public static final TagKey<Item> FOODS_RAW_BACON = commonItemTag("foods/raw_bacon");
     public static final TagKey<Item> FOODS_RAW_BEEF = commonItemTag("foods/raw_beef");